# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/complaints/{complaintId}/attachments": {
      "post": {
        "operationId": "AttachmentsController_upload",
        "parameters": [
          {
            "name": "complaintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaint attachments"
        ]
      },
      "get": {
        "operationId": "AttachmentsController_list",
        "parameters": [
          {
            "name": "complaintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaint attachments"
        ]
      }
    },
    "/complaints/{complaintId}/attachments/{attachmentId}": {
      "get": {
        "operationId": "AttachmentsController_retrieve",
        "parameters": [
          {
            "name": "complaintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "complaint attachments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urbanity API",
    "description": "In-memory NestJS CRUD API with JWT-backed role-based access control.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}